+++ packages/@magic-ext/auth/src/types.ts
@@ -1,0 +1,7 @@
+export interface UserInfo {
+  email?: string;
+}
+
+export interface WalletInfo {
+  walletType: 'magic' | 'metamask' | 'coinbase_wallet' | 'wallet_connect';
+}