+++ packages/@magic-sdk/provider/src/modules/auth.ts
@@ -5,10 +5,20 @@ import {
   LoginWithEmailOTPConfiguration,
   LoginWithEmailOTPEvents,
   LoginWithMagicLinkEvents,
+  UpdateEmailConfiguration,
 } from '@magic-sdk/types';
 import { BaseModule } from './base-module';
 import { createJsonRpcRequestPayload } from '../core/json-rpc';
 import { SDKEnvironment } from '../core/sdk-environment';
+import { UpdateEmailEvents } from './user';
+import { createDeprecationWarning } from '../core/sdk-exceptions';
+
+export const ProductConsolidationMethodRemovalVersions = {
+  'magic-sdk': 'v18.0.0',
+  '@magic-sdk/react-native': 'v14.0.0',
+  '@magic-sdk/react-native-bare': 'v19.0.0',
+  '@magic-sdk/react-native-expo': 'v19.0.0',
+};
 
 export class AuthModule extends BaseModule {
   /**
@@ -17,6 +27,11 @@ export class AuthModule extends BaseModule {
    * of 15 minutes).
    */
   public loginWithMagicLink(configuration: LoginWithMagicLinkConfiguration) {
+    createDeprecationWarning({
+      method: 'auth.loginWithMagicLink()',
+      removalVersions: ProductConsolidationMethodRemovalVersions,
+      useInstead: '@magic-ext/auth auth.loginWithMagicLink()',
+    }).log();
     const { email, showUI = true, redirectURI } = configuration;
 
     const requestPayload = createJsonRpcRequestPayload(
@@ -32,6 +47,11 @@ export class AuthModule extends BaseModule {
    * of 15 minutes)
    */
   public loginWithSMS(configuration: LoginWithSmsConfiguration) {
+    createDeprecationWarning({
+      method: 'auth.loginWithSMS()',
+      removalVersions: ProductConsolidationMethodRemovalVersions,
+      useInstead: '@magic-ext/auth auth.loginWithSMS()',
+    }).log();
     const { phoneNumber } = configuration;
     const requestPayload = createJsonRpcRequestPayload(
       this.sdk.testMode ? MagicPayloadMethod.LoginWithSmsTestMode : MagicPayloadMethod.LoginWithSms,
@@ -46,6 +66,11 @@ export class AuthModule extends BaseModule {
    * of 15 minutes)
    */
   public loginWithEmailOTP(configuration: LoginWithEmailOTPConfiguration) {
+    createDeprecationWarning({
+      method: 'auth.loginWithEmailOTP()',
+      removalVersions: ProductConsolidationMethodRemovalVersions,
+      useInstead: '@magic-ext/auth auth.loginWithEmailOTP()',
+    }).log();
     const { email, showUI } = configuration;
     const requestPayload = createJsonRpcRequestPayload(
       this.sdk.testMode ? MagicPayloadMethod.LoginWithEmailOTPTestMode : MagicPayloadMethod.LoginWithEmailOTP,
@@ -77,6 +102,11 @@ export class AuthModule extends BaseModule {
    * `window.location.search`.
    */
   public loginWithCredential(credentialOrQueryString?: string) {
+    createDeprecationWarning({
+      method: 'auth.loginWithCredential()',
+      removalVersions: ProductConsolidationMethodRemovalVersions,
+      useInstead: '@magic-ext/auth auth.loginWithCredential()',
+    }).log();
     let credentialResolved = credentialOrQueryString ?? '';
 
     if (!credentialOrQueryString && SDKEnvironment.platform === 'web') {
@@ -97,7 +127,38 @@ export class AuthModule extends BaseModule {
 
   // Custom Auth
   public setAuthorizationToken() {
+    createDeprecationWarning({
+      method: 'auth.setAuthorizationToken()',
+      removalVersions: ProductConsolidationMethodRemovalVersions,
+      useInstead: '@magic-ext/auth auth.setAuthorizationToken()',
+    }).log();
     const requestPayload = createJsonRpcRequestPayload(MagicPayloadMethod.SetAuthorizationToken);
     return this.request<boolean>(requestPayload);
+  }
+
+  public updateEmailWithUI(configuration: UpdateEmailConfiguration) {
+    createDeprecationWarning({
+      method: 'auth.updateEmailWithUI()',
+      removalVersions: ProductConsolidationMethodRemovalVersions,
+      useInstead: '@magic-ext/auth auth.updateEmailWithUI()',
+    }).log();
+    const { email, showUI = true } = configuration;
+    const requestPayload = createJsonRpcRequestPayload(
+      this.sdk.testMode ? MagicPayloadMethod.UpdateEmailTestMode : MagicPayloadMethod.UpdateEmail,
+      [{ email, showUI }],
+    );
+    return this.request<string | null, UpdateEmailEvents>(requestPayload);
+  }
+
+  public updatePhoneNumberWithUI() {
+    createDeprecationWarning({
+      method: 'auth.updatePhoneNumberWithUI()',
+      removalVersions: ProductConsolidationMethodRemovalVersions,
+      useInstead: '@magic-ext/auth auth.updatePhoneNumberWithUI()',
+    }).log();
+    const requestPayload = createJsonRpcRequestPayload(
+      this.sdk.testMode ? MagicPayloadMethod.UpdatePhoneNumberTestMode : MagicPayloadMethod.UpdatePhoneNumber,
+    );
+    return this.request<string | null>(requestPayload);
   }
 }